test: Fix for potential wrong sample data for product image zoom e2e test (#14995)
