+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/image-zoom/image-zoom.e2e-spec.ts
@@ -1,23 +1,34 @@
 import { waitForProductPage } from '../../../helpers/checkout-flow';
-import { interceptProductDetails } from '../../../helpers/product-details';
 import { viewportContext } from '../../../helpers/viewport-context';
 
 const productId = '1990255';
 
 describe('Image zoom', () => {
   viewportContext(['desktop'], () => {
+    let hasGallery = false;
+
     before(() => {
       cy.window().then((win) => {
         win.sessionStorage.clear();
       });
       const productPage = waitForProductPage(productId, 'getProductPage');
       cy.visit(`/product/${productId}`);
-      const productDetailsAlias = interceptProductDetails(productId);
+      cy.intercept(
+        'GET',
+        `${Cypress.env('OCC_PREFIX')}/${Cypress.env(
+          'BASE_SITE'
+        )}/products/${productId}?fields=*stock(DEFAULT)*&lang=en&curr=USD`
+      ).as('getProductDetails');
 
       cy.wait(`@${productPage}`).its('response.statusCode').should('eq', 200);
-      cy.wait(`@${productDetailsAlias}`)
-        .its('response.statusCode')
-        .should('eq', 200);
+      cy.wait('@getProductDetails').then((xhr) => {
+        expect(xhr.response.statusCode).to.equal(200);
+
+        // Checking if product has more than 1 image to show gallery
+        hasGallery = xhr.response?.body?.images
+          ?.filter((image) => image.imageType === 'GALLERY')
+          .some((image) => image.galleryIndex > 0);
+      });
     });
 
     it('should display image zoom trigger', () => {
@@ -39,10 +50,6 @@ describe('Image zoom', () => {
       cy.get('cx-media.cx-default-image-zoom').should('be.visible');
     });
 
-    it('thumbnails should be visible', () => {
-      cy.get('cx-product-image-zoom-thumbnails').should('be.visible');
-    });
-
     it('clicking on main image should zoom in image', () => {
       cy.get('cx-media.cx-default-image-zoom').click();
 
@@ -55,42 +62,58 @@ describe('Image zoom', () => {
       cy.get('cx-media.cx-default-image-zoom').should('be.visible');
     });
 
-    it('clicking on thumbnail should open corresponding main image', () => {
-      cy.get('cx-carousel cx-media img')
-        .eq(1)
-        .invoke('prop', 'src')
-        .as('thumbnail1');
-
-      cy.get('cx-carousel cx-media img').eq(1).click();
-
-      cy.get('cx-media.cx-default-image-zoom img')
-        .invoke('prop', 'src')
-        .as('main1');
-
-      cy.get('@thumbnail1').then((thumb1) => {
-        cy.get('@main1').then((main1) => {
-          expect(main1).to.equal(thumb1);
-        });
-      });
+    it('thumbnails should be visible', () => {
+      if (hasGallery) {
+        cy.get('cx-product-image-zoom-thumbnails').should('be.visible');
+      } else {
+        cy.log('Product has only 1 image, not possible to run this test.');
+      }
     });
 
-    it('clicking on navigation arrow should change main image', () => {
-      cy.get('cx-media.cx-default-image-zoom img')
-        .invoke('prop', 'src')
-        .as('mainInitial');
+    it('clicking on thumbnail should open corresponding main image', () => {
+      if (hasGallery) {
+        cy.get('cx-carousel cx-media img')
+          .eq(1)
+          .invoke('prop', 'src')
+          .as('thumbnail1');
 
-      cy.get('@mainInitial').then((mainInitial) => {
-        cy.get('cx-icon.fa-angle-right').first().click();
+        cy.get('cx-carousel cx-media img').eq(1).click();
 
         cy.get('cx-media.cx-default-image-zoom img')
           .invoke('prop', 'src')
-          .as('mainAfter');
+          .as('main1');
 
-        cy.get('@mainAfter').then((mainAfter) => {
-          expect(mainInitial).not.to.equal(mainAfter);
-          cy.get('cx-icon.fa-angle-left').eq(1).click();
+        cy.get('@thumbnail1').then((thumb1) => {
+          cy.get('@main1').then((main1) => {
+            expect(main1).to.equal(thumb1);
+          });
         });
-      });
+      } else {
+        cy.log('Product has only 1 image, not possible to run this test.');
+      }
+    });
+
+    it('clicking on navigation arrow should change main image', () => {
+      if (hasGallery) {
+        cy.get('cx-media.cx-default-image-zoom img')
+          .invoke('prop', 'src')
+          .as('mainInitial');
+
+        cy.get('@mainInitial').then((mainInitial) => {
+          cy.get('cx-icon.fa-angle-right').first().click();
+
+          cy.get('cx-media.cx-default-image-zoom img')
+            .invoke('prop', 'src')
+            .as('mainAfter');
+
+          cy.get('@mainAfter').then((mainAfter) => {
+            expect(mainInitial).not.to.equal(mainAfter);
+            cy.get('cx-icon.fa-angle-left').eq(1).click();
+          });
+        });
+      } else {
+        cy.log('Product has only 1 image, not possible to run this test.');
+      }
     });
   });
 });
